removing useless code and add tests
Removing code to dynamically load libraries, since it will not be really implemented right now, it will just be a pain to maintain.

Adding tests to check printing in all three cases of tests (success, fail, and skip). But for now, suites aren't really tested, and data aren't tested either.