Basic functionality and test runner (#5)
* create a basic directory structure for the project

The following structure will be used:
.
├── CMakeLists.txt
├── include
│   └── CLarity
│       └── clarity.h
└── src
    └── clarity.c

* create a basic directory structure for the project

The following structure will be used:
.
├── CMakeLists.txt
├── include
│   └── CLarity
│       └── clarity.h
└── src
    └── clarity.c

* improve build system

It still needs some improvements regarding tests.

* add dynamic loader API

This API is platform independent, but only windows is implemented for now.
There is also the base for the actual runner.
The test loader is implemented.

* add suite API and its implementation

* add an implementation for the test API

* add the actual suite runner

Everything for now isn't tested, since the actual build system for testing isn't implemented yet.

The runner doesn't make use of processes for now. It is subject to memory issues, and to crashes.

The print functions are subject to improvements and will be reviewed in a future issue.

* 7 implement basic test cases (#8)

* add test architecture

The build infrastructure for testing is set. Now the tests still needs to be implemented.

* small update to the printer and remove useless cache for TEST_TARGETS

* rename the only test for now

test_one_suite_one_test.c -> test_one_suite_one_test_pass.c

* removing useless code and add tests

Removing code to dynamically load libraries, since it will not be really implemented right now, it will just be a pain to maintain.

Adding tests to check printing in all three cases of tests (success, fail, and skip). But for now, suites aren't really tested, and data aren't tested either.

* add tests for suites

Testing multiple suites and multiple tests in a single suite.